# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Clinic {
  id: number;
  name: string;
  specialty: string;
  neighborhood: string;
  address: string;
  phone: string;
  /** Typical days to get an appointment (vs. ~21 days at NYU clinics) */
  averageWaitDays: number;
  acceptsNyuInsurance: boolean;
  rating: number;
  distanceFromCampusMiles: number;
  hours: string;
  notes: string;
  walkInAvailable: boolean;
}

export interface SpecialtyCount {
  specialty: string;
  count: number;
}

export interface ClinicsSummary {
  totalClinics: number;
  averageWaitDays: number;
  walkInCount: number;
  specialtyBreakdown: SpecialtyCount[];
}

export interface InsurancePlan {
  id: number;
  name: string;
  description: string;
  annualPremium: number;
  waivable: boolean;
  keyBenefits: string[];
}

export type ListClinicsParams = {
specialty?: string;
neighborhood?: string;
acceptsNyuInsurance?: boolean;
search?: string;
};