# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List alternative clinics
      description: Returns clinics, optionally filtered by specialty, neighborhood, or NYU insurance acceptance
      parameters:
        - name: specialty
          in: query
          required: false
          schema:
            type: string
        - name: neighborhood
          in: query
          required: false
          schema:
            type: string
        - name: acceptsNyuInsurance
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of clinics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"
  /clinics/{id}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get a single clinic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clinic detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Clinic:
      type: object
      required:
        - id
        - name
        - specialty
        - neighborhood
        - address
        - phone
        - averageWaitDays
        - acceptsNyuInsurance
        - rating
        - distanceFromCampusMiles
        - hours
        - notes
        - walkInAvailable
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        neighborhood:
          type: string
        address:
          type: string
        phone:
          type: string
        averageWaitDays:
          type: number
          description: Typical days to get an appointment (vs. ~21 days at NYU clinics)
        acceptsNyuInsurance:
          type: boolean
        rating:
          type: number
        distanceFromCampusMiles:
          type: number
        hours:
          type: string
        notes:
          type: string
        walkInAvailable:
          type: boolean
...