# lib/api-zod/src/generated/types/clinic.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Clinic {
  id: number;
  name: string;
  specialty: string;
  neighborhood: string;
  address: string;
  phone: string;
  /** Typical days to get an appointment (vs. ~21 days at NYU clinics) */
  averageWaitDays: number;
  acceptsNyuInsurance: boolean;
  rating: number;
  distanceFromCampusMiles: number;
  hours: string;
  notes: string;
  walkInAvailable: boolean;
}